Cambios en operaciones

+++ src/main/java/co/unbosque/isii/calculadora_pruebas/Comando.java
@@ -2,7 +2,7 @@ package co.unbosque.isii.calculadora_pruebas;
 
 public interface Comando {
 
-	double ejecutarOperacion(double d, int i);
+	double ejecutarOperacion(double d, double i);
 
 	String separar(String cadena);
